move e2e tests to the component's src dir
