+++ src/material-experimental/mdc-chips/BUILD.bazel
@@ -1,7 +1,8 @@
 package(default_visibility=["//visibility:public"])
 
 load("@io_bazel_rules_sass//:defs.bzl", "sass_library", "sass_binary")
-load("//tools:defaults.bzl", "ng_module")
+load("@npm_angular_bazel//:index.bzl", "protractor_web_test_suite")
+load("//tools:defaults.bzl", "ng_module", "ts_library")
 
 ng_module(
   name = "mdc-chips",
@@ -36,4 +37,33 @@ sass_binary(
     "//src/material/core:all_themes",
     "//src/material-experimental/mdc-helpers:mdc_scss_deps_lib",
   ]
+)
+
+ts_library(
+  name = "e2e_test_sources",
+  srcs = glob(["**/*.e2e.spec.ts"]),
+  tsconfig = "//src/e2e-app:tsconfig-e2e.json",
+  deps = [
+    "@npm//@types/jasmine",
+    "@npm//@types/selenium-webdriver",
+    "@npm//protractor",
+    "//src/e2e-app:e2e_utils_lib",
+  ]
+)
+
+protractor_web_test_suite(
+  name = "e2e_tests",
+  tags = ["e2e"],
+  configuration = "//src/e2e-app:protractor.conf.js",
+  on_prepare = "//src/e2e-app:start-devserver.js",
+  server = "//src/e2e-app:devserver",
+  deps = [
+    "@npm//protractor",
+    ":e2e_test_sources",
+    "//src/e2e-app:e2e_utils_lib",
+  ],
+  data = [
+    "@npm//@angular/bazel",
+    "//tools/axe-protractor",
+  ],
 )
