+++ src/material/icon/icon.e2e.spec.ts
