+++ src/material/menu/menu.e2e.spec.ts
@@ -5,13 +5,13 @@ import {
   expectLocation,
   expectToExist,
   pressKeys,
-} from '../../src/e2e-app/test-util/index';
+} from '../../e2e-app/test-util';
 
 const presenceOf = ExpectedConditions.presenceOf;
 const not = ExpectedConditions.not;
 
 describe('menu', () => {
-  const menuSelector = '.mat-mdc-menu-panel';
+  const menuSelector = '.mat-menu-panel';
   const page = {
     menu: () => element(by.css(menuSelector)),
     start: () => element(by.id('start')),
@@ -29,7 +29,7 @@ describe('menu', () => {
     getResultText: () => page.textArea().getText(),
   };
 
-  beforeEach(async () => await browser.get('/mdc-menu'));
+  beforeEach(async () => await browser.get('/menu'));
 
   it('should open menu when the trigger is clicked', async () => {
     await expectToExist(menuSelector, false);
@@ -84,7 +84,7 @@ describe('menu', () => {
 
   it('should mirror classes on host to menu template in overlay', async () => {
     await page.trigger().click();
-    expect(await page.menu().getAttribute('class')).toContain('mat-mdc-menu-panel');
+    expect(await page.menu().getAttribute('class')).toContain('mat-menu-panel');
     expect(await page.menu().getAttribute('class')).toContain('custom');
   });
