+++ src/material/progress-bar/progress-bar.e2e.spec.ts
@@ -1,5 +1,5 @@
 import {browser} from 'protractor';
-import {expectToExist} from '../../src/e2e-app/test-util/index';
+import {expectToExist} from '../../e2e-app/test-util';
 
 describe('progress-bar', () => {
   beforeEach(async () => await browser.get('/progress-bar'));
